[feature/formats] check for all netpbm formats, and generic preamble;
We now check for all six format specifiers, though we don't actually change the
behaviour based on it.  That is, we ignore it and just stubbornly output P1
format still.  But this is the preliminary groundwork for supporting those.

There's also a generic preamble function for all six netpbm formats.  It's a
little messy though, and I intend to make the parameter passing a little bit
more compact and readable before too long.