[feature/better-threading] first stab;
Taking a first stab at an improved thread-based backend.  At a minimum,
this is using mutexes.

With the use of condition variables initially, execution flow seemed to
reach a blocking condition, which wasn't clearing.  So I commented-out
the condition variable logic for now.

The run produces an image closely resembling Mandelbrot set, but with
numerous errors.  Points in the set are not output in the proper order.

It's also significantly faster than the first attempt at a thread-based
backend, but its still somewhat slower than the non-threaded backend.

So obviously, this still needs quite a bit more work.