[feature/better-threading] finally works with mutex/cond;
I finally got a more sophiticated use of threads implemented
successfully.  Rather than constantly spawning short-lived threads,
only to have to clean them up moments later, I keep all the threads
around for as long as the program is running, managed in a circular
linked list.

As a thread finishes one work unit, it starts the next available one
in the sequence, with the synchronization handled by the mutex/cond
logic.

This is faster than the simple create/join model, by about a factor of
two, but it's still way slower than the non-threaded model.  I need to
figure-out what is still causing most of the inefficiency.  Perhaps
there's too much overhead in threading to offset the shortness of the
individual work units.  Maybe I need to increase the overall size of
the batch before it begins to break-even and offset the overhead.

Note that this threading scheme definitely has some headroom of its
own, in which it could be more efficient.  Because I'm forcing the
output to happen in a synchronous stream, rather than sending the
output to a buffer in a non-orderd manner, then threads wind-up being
idle more then the should be.  I'm aware of this limitation, but I
just wanted to see how much mileage I could get out of this before I
start allocating large chunks of heap.