[enhancement/long-double] converting from "double" to "long double";
Basically, converting all "double"s in the program to "long double"s,
to squeeze more precision out of it, because I seem to be reaching the
limits of the current precision.

Also changed the conversion specifiers that were being used when parsing
the command-line args and when saving to the image file metadata.

I replaced the floating-point type with a macro, in case I need to
change it again.